+++ CHANGELOG.md
@@ -8,9 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
-- Look up the iamroot libraries in the directory specified in the environment
-  variable `IAMROOT_ORIGIN` if set first and in the iamroot library directory
-  then
+- Look up the iamroot libraries and `exec.sh` in the directory specified in the
+  environment variable `IAMROOT_ORIGIN` if set first and in the iamroot library
+  directory then
 - Set the environment variable `IAMROOT_ORIGIN` using `DT_RUNPATH` if set in
   `ld-iamroot.so(8)`